+++ Cargo.toml
@@ -26,7 +26,6 @@ serde = ["dep:serde"]
 arbitrary = ["dep:arbitrary"]
 generate-bindings = ["dep:bindgen"]
 ethereum_kzg_settings = ["dep:once_cell"]
-custom-allocs = []
 
 # Enable this feature when running the tests to generate the fuzzing corpus.
 # This converts the yaml reference tests into a binary form for the fuzzer.
