+++ bindings/rust/src/zkvm_allocs/mod.rs
